requirements.txt
Pillow
exif